Add neovim lsp
Add omnisharp language server
Add lsp saga
Change plugin manager to packer
Fix wrong window switch keybinding
Change several keybindings to silent
Add floating terminal keybinding
