--- nvim/lang-client.lua
@@ -1,4 +1,0 @@
-require "lspconfig".omnisharp.setup
-{
-    cmd = {  "/usr/bin/omnisharp", "--languageserver", "--hostPID", tostring(vim.fn.getpid()) }
-}
