+++ nvim/lua/plugins.lua
@@ -1,0 +1,9 @@
+return require'packer'.startup(
+    function()
+        use 'wbthomason/packer.nvim'
+
+        use 'neovim/nvim-lspconfig'
+
+        use 'glepnir/lspsaga.nvim'
+    end
+)
