--- nvim/plugin.vim
@@ -1,3 +1,0 @@
-call plug#begin(stdpath('data') . './plugged')
-Plug 'neovim/nvim-lspconfig'
-call plug#end()
